pyproject.toml
[tool.poetry]
name = "tracker"
version = "0.1.0"
description = ""
authors = ["aleksandr <aleksandr_drobnov@epam.com>"]

[tool.poetry.dependencies]
python = "^3.6"
PyTrack-NTU = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
